% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talks20260420PSUHDFSJobtalk-package.R
\docType{package}
\name{talks20260420PSUHDFSJobtalk-package}
\alias{talks20260420PSUHDFSJobtalk}
\alias{talks20260420PSUHDFSJobtalk-package}
\title{talks20260420PSUHDFSJobtalk: Penn State HDFS Job Talk Materials}
\description{
Materials for the Penn State Human Development and Family Studies job talk, delivered on April 20, 2026.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/talks20260420PSUHDFSJobtalk}
  \item \url{https://jeksterslab.github.io/talks20260420PSUHDFSJobtalk/}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]
}

}
\keyword{internal}
